+++ stable_baselines3/common/callbacks.py
@@ -6,7 +6,7 @@ from typing import Any, Callable, Dict, List, Optional, Union
 import gym
 import numpy as np
 
-from stable_baselines3.common import base_class, logger  # pytype: disable=pyi-error
+from stable_baselines3.common import base_class  # pytype: disable=pyi-error
 from stable_baselines3.common.evaluation import evaluate_policy
 from stable_baselines3.common.vec_env import DummyVecEnv, VecEnv, sync_envs_normalization
 
@@ -44,7 +44,7 @@ class BaseCallback(ABC):
         """
         self.model = model
         self.training_env = model.get_env()
-        self.logger = logger
+        self.logger = model.logger
         self._init_callback()
 
     def _init_callback(self) -> None: